splits application_settings.py
